tests/analytics/requirements-test.txt
# Requirements para tests del módulo Analytics
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
psutil>=5.9.0
locust>=2.15.0
requests>=2.28.0